Can add recipes from admin-cp

+++ includes/app/routes/admin-cp.php
@@ -30,33 +30,52 @@ $app->get('/admin-cp', function(Request $request, Response $response)
 
 $app->post('/admin-cp', function(Request $request, Response $response) {
 
-    $db = new UserDatabaseWrapper();
+    if ($_SESSION["isAdmin"] === TRUE) {
 
-    if (isset($_POST['makeAdminSubmit'])) {
-        $makeAdminResult = $db->makeAdmin();
+        $db = new UserDatabaseWrapper();
 
-        if ($makeAdminResult === true) {
-            $arr["success"] = "You have successfully given that account administrative privileges.";
-            $arr["firstName"] = $_SESSION['firstName'];
-            return $this->view->render($response, 'admin-cp-success.html.twig', $arr);
-        } else {
-            $arr["error"] = $makeAdminResult;
-            $arr["firstName"] = $_SESSION['firstName'];
-            return $this->view->render($response, 'admin-cp-error.html.twig', $arr);
+        if (isset($_POST['makeAdminSubmit'])) {
+            $makeAdminResult = $db->makeAdmin();
+
+            if ($makeAdminResult === true) {
+                $arr["success"] = "You have successfully given that account administrative privileges.";
+                $arr["firstName"] = $_SESSION['firstName'];
+                return $this->view->render($response, 'admin-cp-success.html.twig', $arr);
+            } else {
+                $arr["error"] = $makeAdminResult;
+                $arr["firstName"] = $_SESSION['firstName'];
+                return $this->view->render($response, 'admin-cp-error.html.twig', $arr);
+            }
+        }
+
+        if (isset($_POST['removeAdminSubmit'])) {
+            $removeAdminResult = $db->removeAdmin();
+
+            if ($removeAdminResult === true) {
+                $arr["success"] = "You have successfully removed that account's administrative privileges.";
+                $arr["firstName"] = $_SESSION['firstName'];
+                return $this->view->render($response, 'admin-cp-success.html.twig', $arr);
+            } else {
+                $arr["error"] = $removeAdminResult;
+                $arr["firstName"] = $_SESSION['firstName'];
+                return $this->view->render($response, 'admin-cp-error.html.twig', $arr);
+            }
+        }
+
+        if (isset($_POST['recipeSubmit'])) {
+            $recipeSubmitResult = $db->recipeSubmit();
+
+            if ($recipeSubmitResult === true) {
+                $arr["success"] = "You have successfully added that recipe to the database.";
+                $arr["firstName"] = $_SESSION['firstName'];
+                return $this->view->render($response, 'admin-cp-success.html.twig', $arr);
+            } else {
+                $arr["error"] = $recipeSubmitResult;
+                $arr["firstName"] = $_SESSION['firstName'];
+                return $this->view->render($response, 'admin-cp-error.html.twig', $arr);
+            }
         }
     }
 
-    if (isset($_POST['removeAdminSubmit'])) {
-        $removeAdminResult = $db->removeAdmin();
 
-        if ($removeAdminResult === true) {
-            $arr["success"] = "You have successfully removed that account's administrative privileges.";
-            $arr["firstName"] = $_SESSION['firstName'];
-            return $this->view->render($response, 'admin-cp-success.html.twig', $arr);
-        } else {
-            $arr["error"] = $removeAdminResult;
-            $arr["firstName"] = $_SESSION['firstName'];
-            return $this->view->render($response, 'admin-cp-error.html.twig', $arr);
-        }
-    }
 });
